mlops/reporting/requirements-report.txt
openai>=1.13.3
python-dotenv
pandas
requests
